🎨 refactor: change `sources` state type from `Source[]` to `any[]` in `ScreenSources` and `ScreenSourcesV1` 🔥 remove: unused `QuickLinkV1` component
The type of `sources` state in the `ScreenSources` and `ScreenSourcesV1` components was changed from `Source[]` to `any[]`. This change provides more flexibility as it removes the requirement for the sources data to strictly follow the `Source` type structure.

The `QuickLinkV1` component was removed as it was no longer being used in the application. This helps to clean up the code base and improve maintainability.