🖥️ feat: Enhance device identification and registration in Electron app
Improved the device identification process in the Electron app. Integrated a new function, `getComputerName`, to accurately fetch the computer's name based on the operating system. This function uses different methods for Windows, macOS, and Linux, ensuring that the device name is correctly identified regardless of the platform. This enhancement provides more accurate and reliable device identification, which is crucial for features like device management and user sessions.

Extended the device registration URL to include `appVersion` and `deviceType`. When opening a new device registration page, the URL now carries additional parameters: the application version and the type of the operating system. This change allows the server to receive more detailed information about the device, improving device management capabilities and providing a better overview of the user's environment.

Removed unnecessary console logs to clean up the code. The refactoring includes removing logs that were used for debugging but are no longer necessary, contributing to cleaner and more maintainable code. This cleanup reduces clutter in the console and makes it easier to identify important log messages during development and troubleshooting.

These changes collectively enhance the functionality of the Electron app, particularly in how it interacts with device management features, providing a more seamless and informative user experience.