🚀 feat: add speed-insights to measure web performance
Add `@vercel/speed-insights` package to measure web performance metrics and improve user experience. This package is added as a dependency in the `package.json` file and is imported and used in the `layout.tsx` file to track performance metrics.

This change allows us to monitor and optimize the performance of our web application.